% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwsR-package.R
\docType{package}
\name{dwsR-package}
\alias{dwsR}
\alias{dwsR-package}
\title{dwsR: Provide Fast Convolution of Matrixes of Numbers}
\description{
Takes a matrix of numeric data and a kernel, 
  and convolves the data using the matrix.
}
\keyword{internal}
