% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_exponential.R
\name{exponential_kernel}
\alias{exponential_kernel}
\title{Compute an Exponential kernel}
\usage{
exponential_kernel(beta = 0.2, r0 = 0.05)
}
\arguments{
\item{beta}{\strong{[numeric]} The kernel's beta parameter.}

\item{r0}{\strong{[numeric]} The kernel's r0 (exponential).}
}
\value{
A \code{matrix} corresponding to the kernel.
}
\description{
Functions to compute an exponential kernel.
}
