% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-utils.R
\name{kernel_flip_horizontal}
\alias{kernel_flip_horizontal}
\alias{kernel_flip_vertical}
\alias{kernel_flip_both}
\title{Flip a kernel}
\usage{
kernel_flip_horizontal(k)

kernel_flip_vertical(k)

kernel_flip_both(k)
}
\arguments{
\item{k}{\strong{[matrix]} The kernel.}
}
\value{
The flipped kernel, a \code{matrix}.
}
\description{
Flip the kernel vertically, horizontally, or both.
}
\examples{

kernel <- matrix(c(1,2,3,
                   0,1,0,
                   2,3,4), nrow = 3, ncol= 3)
kernel_flip_horizontal(kernel)
kernel_flip_vertical(kernel)
kernel_flip_both(kernel)

}
