% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfocal_info.R
\name{pfocal_transform_info}
\alias{pfocal_transform_info}
\alias{pfocal_reduce_info}
\alias{pfocal_nan_policy_info}
\alias{pfocal_mean_divisor_info}
\alias{pfocal_variance_info}
\title{Retrieve kernel and arguments information}
\usage{
pfocal_transform_info()

pfocal_reduce_info()

pfocal_nan_policy_info()

pfocal_mean_divisor_info()

pfocal_variance_info()
}
\value{
An object of type \code{matrix}: first column shows the possible argument value,
second column is used internally and third column describe the argument.
}
\description{
Different kernels and arguments can be passed to \link{pfocal}. A set of functions
is available to provide all possible values for these arguments.
}
\examples{

# Retrieve info on different arguments

pfocal_transform_info()
pfocal_reduce_info()
pfocal_nan_policy_info()
pfocal_mean_divisor_info()
pfocal_variance_info()

# Higher level, argument level functions calling routine below -------------
}
