% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfocal-package.R
\docType{package}
\name{pfocal-package}
\alias{pfocal-package}
\alias{_PACKAGE}
\title{pfocal: Fast Convolution of Matrixes of Numbers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions to create and apply convolution kernels to rectangular data (matrix, raster). Implements calculations of convolutions in parallel, emphasizing performance over convolution function diversity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://landscitech.github.io/pfocal/}
  \item Report bugs at \url{https://github.com/LandSciTech/pfocal/}
}

}
\author{
\strong{Maintainer}: Josie Hughes \email{josie.hughes@canada.ca}

Authors:
\itemize{
  \item Griffin Barrett \email{griffin-barrett@outlook.com}
  \item Valentin Lucet \email{valentin.lucet@gmail.com}
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Environment and Climate Change [copyright holder]
}

}
\keyword{internal}
