% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFocal.R
\name{pFocal}
\alias{pFocal}
\title{Fast, parallel implementation of raster focal calculations}
\usage{
pFocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)
}
\arguments{
\item{data}{\strong{[matrix-type]} Grid to compute onto.}

\item{kernel}{\strong{[matrix]} Computation kernel (neighborhood).}

\item{edge_value}{\strong{[numeric]} TBD, default to 0.}

\item{transform_function}{\strong{[character]} TBD, default to "MULTIPLY".}

\item{reduce_function}{\strong{[character]} TBD, default to "SUM".}

\item{mean_divider}{\strong{[character]} TBD, default to "ONE".}

\item{variance}{\strong{[logical]} TBD, default to FALSE.}

\item{na.rm}{\strong{[NA OR logical]} TBD, default to NA.}

\item{mp}{\strong{[logical]} TBD, default to TRUE.}

\item{debug_use_r_implementation}{\strong{[logical]} TDB, default to FALSE.}

\item{...}{TDB.}
}
\value{
The updated matrix-type object.
}
\description{
A fast and pluralized implementation of focal and convolutionnal calculation
on raster-type data (matrices and other grid based objects).
}
\examples{

data <- matrix(nrow = 10, ncol = 10, data = runif(10*10))
kernel <- matrix(1/9,nrow=3,ncol=3)
pFocal(data = data, kernel = kernel)

}
