% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwsR-package.R
\docType{package}
\name{dwsR-package}
\alias{dwsR}
\alias{dwsR-package}
\title{dwsR: Fast Convolution of Matrixes of Numbers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Takes a matrix of numeric data and a kernel, 
    and convolves the data using the matrix.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://landscitech.github.io/dwsR}
  \item Report bugs at \url{https://github.com/LandSciTech/dwsR}
}

}
\author{
\strong{Maintainer}: Griffin Barrett \email{griffin-barrett@outlook.com}

Authors:
\itemize{
  \item Josie Hughes \email{josie.hughes@canada.ca}
  \item Valentin Lucet \email{valentin.lucet@gmail.com}
}

}
\keyword{internal}
