% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_binomial.R
\name{binomial_kernel}
\alias{binomial_kernel}
\alias{kernel-binomial}
\title{Compute an Binomial kernel}
\usage{
binomial_kernel(vertical_radius, horizontal_radius = 0)
}
\arguments{
\item{vertical_radius}{\strong{[numeric]} The kernel's radius in the vertical
dimension.}

\item{horizontal_radius}{\strong{[numeric]} The kernel's radius in the horizontal
dimension.}
}
\value{
A \code{matrix} corresponding to the kernel.
}
\description{
Functions to compute a binomial kernel.
}
\examples{

binomial_kernel(vertical_radius = 2, horizontal_radius = 2)

}
