% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfocal-package.R
\docType{package}
\name{pfocal-package}
\alias{pfocal-package}
\alias{_PACKAGE}
\title{pfocal: Fast Convolution of Matrices}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions to create and apply convolution kernels to rectangular data (matrix, raster). Implements calculations of convolutions in parallel, emphasizing performance over convolution function diversity. Moving window calculations are common in many fields, from image processing to raster data manipulation. The process of ingesting values in the neighborhood of a each cell in a grid, passing them through a function, and returning a reduced value, is embarrassingly parallel. This package implements this with parallel C code. In doing so, it provides a much faster option than other similarly aimed packages and functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://landscitech.github.io/pfocal/}
  \item Report bugs at \url{https://github.com/LandSciTech/pfocal/}
}

}
\author{
\strong{Maintainer}: Josie Hughes \email{josie.hughes@ec.gc.ca}

Authors:
\itemize{
  \item Griffin Barrett \email{griffin-barrett@outlook.com}
  \item Valentin Lucet \email{valentin.lucet@gmail.com}
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Environment and Climate Change [copyright holder]
}

}
\keyword{internal}
